@@ -6,32 +6,33 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_spinkit: ^5.1.0
-  flutter_svg: ^0.23.0+1
-  font_awesome_flutter: ^9.2.0
-  google_fonts: ^2.1.0
-  http: ^0.13.0
-  firebase_core: "^1.0.2"
-  firebase_analytics: "^7.1.1"
-  firebase_auth: "^1.0.1"
-  google_sign_in: "^5.0.1"
-  sign_in_with_apple: "^3.0.0"
+  flutter_svg: ^2.0.7
+  font_awesome_flutter: ^10.6.0
+  google_fonts: ^6.1.0
+  http: ^1.1.0
+  firebase_core: ^2.19.0
+  firebase_analytics: ^10.6.1
+  firebase_auth: ^4.11.1
+  google_sign_in: ^6.1.5
+  sign_in_with_apple: ^5.0.0
   crypto: "^3.0.0"
-  cloud_firestore: "^1.0.3"
-  json_annotation: ^4.0.1
-  firebase_messaging: "^9.1.0"
-  firebase_in_app_messaging: ^0.4.0
+  cloud_firestore: ^4.11.0
+  json_annotation: ^4.8.1
+  firebase_messaging: ^14.7.1
+  firebase_in_app_messaging: ^0.7.4+1
+  freezed: ^2.4.5
+  freezed_annotation: ^2.4.1
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.2
-  json_serializable: ^4.1.0
-  build_runner:
+  flutter_launcher_icons: ^0.13.1
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.6
   flutter_test:
     sdk: flutter
 
