@@ -1,5 +1,4 @@
 // ignore: import_of_legacy_library_into_null_safe
-import 'package:firebase_analytics/observer.dart';
 import 'package:flutter/material.dart';
 import 'package:wiredbrain/services/services.dart';
 
@@ -8,8 +7,6 @@ import './get_theme.dart';
 import 'screens/splash_screen.dart';
 
 class CoffeeApp extends StatelessWidget {
-  final FirebaseAnalyticsObserver observer = AnalyticsService.observer;
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -17,7 +14,9 @@ class CoffeeApp extends StatelessWidget {
       themeMode: ThemeMode.system,
       home: const SplashScreen(),
       navigatorKey: CoffeeRouter.instance.navigatorKey,
-      navigatorObservers: <NavigatorObserver>[observer],
+      navigatorObservers: <NavigatorObserver>[
+        AnalyticsService.observer,
+      ],
       theme: getTheme(),
     );
   }
