@@ -6,54 +6,29 @@ part of 'cart_item.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-CartItem _$CartItemFromJson(Map<String, dynamic> json) {
-  return CartItem(
-    id: json['id'] as String?,
-    coffee: Coffee.fromJson(json['coffee'] as Map<String, dynamic>),
-    size: _$enumDecode(_$CoffeeCupSizeEnumMap, json['size']),
-    quantity: json['quantity'] as int,
-    sugar: _$enumDecode(_$CoffeeSugarCubeEnumMap, json['sugar']),
-    additions: (json['additions'] as List<dynamic>)
-        .map((e) => _$enumDecode(_$CoffeeAdditionEnumMap, e))
-        .toList(),
-  );
-}
-
-Map<String, dynamic> _$CartItemToJson(CartItem instance) => <String, dynamic>{
-      'id': instance.id,
-      'coffee': instance.coffee.toJson(),
-      'size': _$CoffeeCupSizeEnumMap[instance.size],
-      'sugar': _$CoffeeSugarCubeEnumMap[instance.sugar],
-      'quantity': instance.quantity,
-      'additions':
-          instance.additions.map((e) => _$CoffeeAdditionEnumMap[e]).toList(),
-    };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
+_$CartItemImpl _$$CartItemImplFromJson(Map<String, dynamic> json) =>
+    _$CartItemImpl(
+      id: json['id'] as String?,
+      coffee: Coffee.fromJson(json['coffee'] as Map<String, dynamic>),
+      size: $enumDecode(_$CoffeeCupSizeEnumMap, json['size']),
+      sugar: $enumDecode(_$CoffeeSugarCubeEnumMap, json['sugar']),
+      quantity: json['quantity'] as int,
+      additions: (json['additions'] as List<dynamic>)
+          .map((e) => $enumDecode(_$CoffeeAdditionEnumEnumMap, e))
+          .toList(),
     );
-  }
 
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
+Map<String, dynamic> _$$CartItemImplToJson(_$CartItemImpl instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'coffee': instance.coffee,
+      'size': _$CoffeeCupSizeEnumMap[instance.size]!,
+      'sugar': _$CoffeeSugarCubeEnumMap[instance.sugar]!,
+      'quantity': instance.quantity,
+      'additions': instance.additions
+          .map((e) => _$CoffeeAdditionEnumEnumMap[e]!)
+          .toList(),
+    };
 
 const _$CoffeeCupSizeEnumMap = {
   CoffeeCupSize.small: 'small',
@@ -67,8 +42,8 @@ const _$CoffeeSugarCubeEnumMap = {
   CoffeeSugarCube.two: 'two',
 };
 
-const _$CoffeeAdditionEnumMap = {
-  CoffeeAddition.cake: 'cake',
-  CoffeeAddition.icecream: 'icecream',
-  CoffeeAddition.cheese: 'cheese',
+const _$CoffeeAdditionEnumEnumMap = {
+  CoffeeAdditionEnum.cake: 'cake',
+  CoffeeAdditionEnum.icecream: 'icecream',
+  CoffeeAdditionEnum.cheese: 'cheese',
 };
