@@ -6,53 +6,29 @@ part of 'order.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Order _$OrderFromJson(Map<String, dynamic> json) {
-  return Order(
-    items: (json['items'] as List<dynamic>)
-        .map((e) => CartItem.fromJson(e as Map<String, dynamic>))
-        .toList(),
-    id: json['id'] as String?,
-    status: _$enumDecode(_$OrderStatusEnumMap, json['status']),
-    userId: json['userId'] as String,
-    created: Order._fromJson(json['created']),
-    updated: Order._fromJson(json['updated']),
-  );
-}
-
-Map<String, dynamic> _$OrderToJson(Order instance) => <String, dynamic>{
-      'items': instance.items.map((e) => e.toJson()).toList(),
-      'status': _$OrderStatusEnumMap[instance.status],
-      'id': instance.id,
-      'userId': instance.userId,
-      'created': Order._toJson(instance.created),
-      'updated': Order._toJson(instance.updated),
-    };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
+_$OrderItemImpl _$$OrderItemImplFromJson(Map<String, dynamic> json) =>
+    _$OrderItemImpl(
+      id: json['id'] as String?,
+      items: (json['items'] as List<dynamic>)
+          .map((e) => CartItem.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      status: $enumDecode(_$OrderStatusEnumMap, json['status']),
+      userId: json['userId'] as String,
+      created:
+          const DateTimeConverter().fromJson(json['created'] as FieldValue),
+      updated:
+          const DateTimeConverter().fromJson(json['updated'] as FieldValue),
     );
-  }
 
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
+Map<String, dynamic> _$$OrderItemImplToJson(_$OrderItemImpl instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'items': instance.items,
+      'status': _$OrderStatusEnumMap[instance.status]!,
+      'userId': instance.userId,
+      'created': const DateTimeConverter().toJson(instance.created),
+      'updated': const DateTimeConverter().toJson(instance.updated),
+    };
 
 const _$OrderStatusEnumMap = {
   OrderStatus.pending: 'pending',
