@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:wiredbrain/enums/enums.dart';
-import 'package:wiredbrain/services/services.dart';
 
 import '../coffee_router.dart';
+import '../enums/enums.dart';
+import '../services/services.dart';
 import 'home.dart';
 import 'menu.dart';
 
@@ -53,7 +53,7 @@ class _SplashScreenState extends State<SplashScreen> {
       () {
         if (_authService.currentUser != null) {
           final String userId = _authService.currentUser!.uid;
-          
+
           _firestoreService.setUserLastLoginTimestamp(userId);
 
           _firestoreService.addLog(
