@@ -1,4 +1,3 @@
-// ignore: import_of_legacy_library_into_null_safe
 import 'package:firebase_in_app_messaging/firebase_in_app_messaging.dart';
 
 class InAppMessagingService {
@@ -11,9 +10,9 @@ class InAppMessagingService {
 
   static InAppMessagingService get instance => _service;
 
-  // triger an Analytics events
-  triggerEvent(String event) {
-    _firebaseInAppMessaging.triggerEvent(event);
+  // trigger an Analytics events
+  Future<void> triggerEvent(String event) async {
+    return _firebaseInAppMessaging.triggerEvent(event);
   }
 
   /// Enables or disables suppression of message displays.
